requirements.txt
Django==4.2.7
aiogram==3.3.0
python-dotenv==1.0.0
Pillow==10.1.0
aiofiles==23.2.1
aiohttp==3.9.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1